ADD: logic for directory transversal, cleanup code

+++ convert.ps1
@@ -1,11 +1,11 @@
 ## Author:Brycen Medart
 ## URL:https://treantlabs.com
-## Version: 1.1
+## Version: 2.0
 ## License: MIT License
 
-#   Argument Variables - Do not change unless use case is understood
-#   Variables for path, category and full file path arguments
- 
+# Argument Variables - Do not change unless use case is understood
+# Variables for path, category and full file path arguments
+
 $inputObj = (
     @{
         category  = $Args[0]
@@ -13,43 +13,42 @@ $inputObj = (
     }
 )
 
-#   Global change variables - Change to configure for your environment.
+# Global change variables - Change to configure for your environment.
 
-#   Filetype array
+# Filetype array
 $ext_array = @(".avi", ".flv", ".mkv", ".mov", ".mp4", ".m4v", ".wmv", ".srt", ".en.srt")
 
-#   FFMPEG variables
+# FFMPEG variables
 $crf_quality = 24
-$preset = "medium"
+$preset = "fast"
 $desired_acodec = "aac"
 
-#   This setting will remove non MP4 files after processing and remove non video files in folder | 0=no 1=yes
-$del_file = 1
+# This setting will remove non MP4 files after processing and remove non video files in folder | 0=no 1=yes
+$delete_files = 0
 
-#   Remove parent directory upon conversion completion | 0=no 1=yes
-$del_parent = 1
+# Remove parent directory upon conversion completion | 0=no 1=yes
+$delete_parent_folder = 0
 
-#   Enable SFTP Transfer | 0=no 1=yes
+# Enable SFTP Transfer | 0=no 1=yes
 $sftp_transfer = 1
 
-#   FFMPEG and WINSCP folders
+# FFMPEG and WINSCP folders
 $ffmpeg_location = "D:\tor\executables\ffmpeg\bin"
 $winscp_location = "D:\tor\executables\WinSCP-5.21.2-Automation"
 
-#   Remote Server Information For SFTP Transfer. Alter these values to match your environment.
-$ssh_fingerprint = "ssh-ed25519 255 3WFfI1chFGA81IBcr3wsGOL6qDopBxvSzdRv+WJUqbY"
-$remote_address = "192.168.1.119"
-$remote_user = "brymed"
-$remote_password = "Brymed#2022"
-$remote_folder = "/media/Videos"
+# Remote Server Information For SFTP Transfer. Alter these values to match your environment.
+$ssh_fingerprint = ""
+$remote_address = ""
+$remote_user = ""
+$remote_password = ""
+$remote_folder = ""
 
 function transfer_file($file) {
 
     try {
         # Load WinSCP .NET assembly
         Add-Type -Path "$($winscp_location)\WinSCPnet.dll"
- 
-        Write-Host "Info: Transferring file to remote server"
+        Write-Host "Info: Transferring $($file) to remote server"
 
         # Set up session options
         $sessionOptions = New-Object WinSCP.SessionOptions -Property @{
@@ -79,15 +78,17 @@ function transfer_file($file) {
             # Print results
             foreach ($transfer in $transferResult.Transfers) {
                 Write-Host "Info: Upload succeeded"
-                Write-Host "Warn: Removing file after transfer"
-                Remove-Item -LiteralPath $file
+
+                if ($delete_files -eq 1) {
+                    Write-Host "Warn: Removing local file after transfer"
+                    Remove-Item -LiteralPath $file
+                }
             }
         }
         finally {
             # Disconnect, clean up
             $session.Dispose()
         }
- 
     }
     catch {
         Write-Host "Error: $($_.Exception.Message)"
@@ -101,61 +102,69 @@ function ConvertFile($path) {
    
     if ($file.Extension -ne ".srt" -And $file.Extension -ne ".en.srt") {
 
-        #   Assign Video/Audio codecs to variables
+        # Assign Video/Audio codecs to variables
         $ffprobe = Get-Command -CommandType Application -Name "$($ffmpeg_location)\ffprobe.exe" -ErrorAction SilentlyContinue
         $ffprobe = $ffprobe.Source
         $ffmpeg = Get-Command -CommandType Application -Name "$($ffmpeg_location)\ffmpeg.exe" -ErrorAction SilentlyContinue
         $ffmpeg = $ffmpeg.Source
 
-        #$vcodec = $(& $ffprobe -v error -select_streams v:0 -show_entries stream=codec_name -of default=nokey=1:noprint_wrappers=1 $file.FullName)
-
-        #convert to reduce file size no matter the extension
+        # Convert to reduce file size no matter the extension
+        Write-Host " "
         Write-Host "Info: Beginning conversion of - $($file.FullName)"
 
         $tmp_file = "$($file.DirectoryName)\tmp_$($file.BaseName).mp4"
 
-        & $ffmpeg -v quiet -stats -i $file.FullName -c:v libx264 -preset $($preset) -crf $($crf_quality) -c:a $($desired_acodec) $tmp_file
-        Write-Host "Info: File Conversion complete!"
-                    
-        #   delete original file after conversion
-        Remove-Item  -LiteralPath $file.FullName
+        # Convert file using ffmpeg and argument variables
+        # Format: video codec: H264
 
-        #   Rename converted file to the original file name that is now deleted
+        & $ffmpeg -v quiet -stats -i $file.FullName -c:v libx264 -preset $($preset) -crf $($crf_quality) -c:a $($desired_acodec) $tmp_file
+ 
+        Write-Host "Info: File Conversion complete!"
+        
+        if ($delete_files -eq 1) {
+            # Delete original file after conversion
+            Remove-Item  -LiteralPath $file.FullName
+        }
+        else {
+            # Rename original file after conversion
+            Rename-item -LiteralPath $file -NewName "old.$($file.basename)$($file.Extension)"
+        }
+    
+        # Rename converted file to the original file name that is now deleted/renamed
         Rename-Item -LiteralPath $tmp_file -NewName "$($file.BaseName).mp4"
 
-        #   If sftp_transfer enabled pass to sftp_transfer function
+        # Transfer if enabled
         if ($sftp_transfer -eq 1) {
             transfer_file("$($file.DirectoryName)\$($file.BaseName).mp4")
         }
-                    
     }
     else {
-        #   Rename srt file with en extension
+        # Rename srt file with .en extension
         Rename-Item -LiteralPath $file -NewName "$($file.BaseName).en.srt"
                 
         if ($sftp_transfer -eq 1) {
+            Write-Host " "
             transfer_file("$($file.DirectoryName)\$($file.BaseName).en.srt")
         }
     }
 
 }
 
-
-#   Identify files before sending to ConvertFile function for video conversion
-
 Write-Host @"
 Author: Brycen Medart 
 URL: https://treantlabs.com
-Version: 1.0
+Version: 2.0
 License: MIT License
 -----------------------------
 
 "@
 
-#   If path supplied is file
+# Identify files before sending to ConvertFile function for video conversion
+
+# If path supplied is file
 if ((Test-Path -LiteralPath $inputObj.full_path -PathType leaf) -eq $true) {
     
-    #   Write-Host "File Provided"
+    Write-Host "File Provided: $($inputObj.FullName)"
     
     $file = Get-Item -LiteralPath $inputObj.full_path
 
@@ -165,16 +174,14 @@ if ((Test-Path -LiteralPath $inputObj.full_path -PathType leaf) -eq $true) {
 
 }
 
-#   Path supplied is directory
+# Path supplied is directory
 else {
-    
     
     Write-Host "Info: Directory Provided -" $inputObj.full_path
 
     $files = Get-ChildItem -LiteralPath $inputObj.full_path
     
     foreach ($file in $files) {
-        # Write-Host $file
 
         #if sub-directory transverse.
         if ((Test-Path -Path $file.FullName -PathType Container) -eq $true) {
@@ -183,38 +190,39 @@ else {
             $subfile = Get-ChildItem -LiteralPath $file.FullName
 
             foreach ($sub in $subfile) {
+
                 if ($sub.Extension -in $ext_array) {
                     ConvertFile($sub.FullName)
                 }
                 else {
-                    if ($del_file -eq 1) {
+                    if ($delete_files -eq 1) {
                         Write-Host 'Warn:' $sub.name "- not present in the extention array - removing..."
                         Remove-Item  -LiteralPath $sub.FullName
                     }
                 }
+
             }
         }
         else {
+
             if ($file.Extension -in $ext_array) {
                 ConvertFile($file.FullName)
             }
-            
             else {
-                if ($del_file -eq 1) {
+                if ($delete_files -eq 1) {
                     Write-Host 'Warn:' $file.name "- not present in the extention array - removing..."
                     Remove-Item  -LiteralPath $file.FullName
                 }
 
             }
+
         }
-
-     
-
     }
 }
 
 #if directory empty and exists, remove
-if ($del_parent -eq 1) {
+if ($delete_parent_folder -eq 1) {
+
     if (((Test-Path -Path "$($inputObj.full_path)\*") -eq $False) -AND (Test-Path -Path "$($inputObj.full_path)") -eq $True) {
         Write-Host "Info: Directory empty, removing folder - $($inputObj.full_path)"
         Remove-Item -LiteralPath $inputObj.full_path
@@ -223,6 +231,7 @@ if ($del_parent -eq 1) {
     else {
         Write-Host "Warn: Parent folder not empty, skipping removal"
     }   
+
 }
 
 Write-Host -NoNewLine 'Press any key to continue...';
